+++ client/src/components/Ratings_Reviews/Ratings.module.css
@@ -238,6 +238,12 @@
   display: flex;
 }
 
+.imagebutton{
+  display: none;
+  cursor: pointer;
+  border: none;
+  background: none;
+}
 .blueText {
   color: blue;
 }
