+++ client/src/components/Ratings_Reviews/StatsTab/RatingBars.jsx
@@ -21,10 +21,11 @@ const bars = (obj) => {
             {i}
           </div>
           <div className={styles.barBar}>
-            <div style = {{ background: 'green',
-            width: `${percentage}%`,
-            height: '100%',
-            }}> </div>
+            <div style={{ background: 'green',
+              width: `${percentage}%`,
+              height: '100%',
+            }}
+            > </div>
 
           </div>
         </div>
